OFFENCES-119 Ignore any directly passed child (inchoate) offences for linking/unlinking to schedules. Plus use schedule ref data in tests + some refactoring

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/manageoffencesapi/service/ScheduleService.kt
@@ -39,14 +39,20 @@ class ScheduleService(
     }
   }
 
-  //  If the offence has any children (inchoate offences) they are linked as well
+  /*
+    If an offence in @offenceIds has any children (inchoate offences) they are linked automatically
+    Any offences in @offenceIds that are themselves children will be ignored
+   */
   @Transactional
   fun linkOffences(schedulePartId: Long, offenceIds: List<Long>) {
     val schedulePart = schedulePartRepository.findById(schedulePartId)
       .orElseThrow { EntityNotFoundException("No schedulePart exists for $schedulePartId") }
 
     val childOffenceIds = offenceRepository.findByParentOffenceIdIn(offenceIds).map { it.id }
-    val offences = offenceRepository.findAllById(offenceIds.plus(childOffenceIds).distinct())
+
+    val offences = offenceRepository
+      .findAllById(offenceIds.plus(childOffenceIds).distinct())
+      .filter { it.parentCode == null || childOffenceIds.contains(it.id) }
 
     offenceSchedulePartRepository.saveAll(
       offences.map { offence ->
@@ -58,10 +64,19 @@ class ScheduleService(
     )
   }
 
-  //  If the offence has any children (inchoate offences) they are unlinked as well
+  /*
+    If an offence in @offenceSchedulePartIds has any children (inchoate offences) they are unlinked automatically
+    Any offences in @offenceIds that are themselves children will be ignored
+   */
   @Transactional
   fun unlinkOffences(offenceSchedulePartIds: List<SchedulePartIdAndOffenceId>) {
+
+    val allOffenceIds = offenceSchedulePartIds.map { it.offenceId }
+    val parentOffenceIds = offenceRepository.findAllById(allOffenceIds).filter { it.parentCode == null }.map { it.id }
+
     offenceSchedulePartIds.forEach {
+      if (!parentOffenceIds.contains(it.offenceId)) return@forEach // ignore any children that have been directly passed in
+
       val childOffenceIds = offenceRepository.findByParentOffenceId(it.offenceId).map { child -> child.id }
       offenceSchedulePartRepository.deleteBySchedulePartIdAndOffenceId(
         it.schedulePartId,
